+++ Models/OrderItem.cs
@@ -4,7 +4,7 @@ namespace FoodDeliveryAPI.Models
     {
         public int orderItemId { get; set; }
         public int? orderItemQuantity { get; set; }
-        public double? orderItemPrice { get; set; }
+        public decimal? orderItemPrice { get; set; }
         public int? orderId { get; set; }
         public int? itemInformationId { get; set; }
     }
